Update to 1.6.1
https://github.com/JoeAlisson/L2jOrg/releases/tag/r1.6.1

+++ AuthServer/src/main/org.l2j.authserver/org/l2j/authserver/network/client/packet/auth2client/LoginOk.java
@@ -21,12 +21,15 @@ package org.l2j.authserver.network.client.packet.auth2client;
 import org.l2j.authserver.network.client.AuthClient;
 import org.l2j.authserver.network.client.packet.AuthServerPacket;
 
+/**
+ * Format: dddddddd f: the session key d: ? d: ? d: ? d: ? d: ? d: ? b: 16 bytes - unknown
+ */
 public final class LoginOk extends AuthServerPacket {
 
 	@Override
 	protected void writeImpl(AuthClient client) {
 		var sessionKey = client.getSessionKey();
-		writeByte(0x03);
+		writeByte((byte)0x03);
 		writeInt(sessionKey.getAuthAccountId());
 		writeInt(sessionKey.getAuthKey());
 		writeBytes(new byte[8]);
