+++ Gameserver/src/main/org.l2j.gameserver/org/l2j/gameserver/network/authcomm/gs2as/PlayerAuthRequest.java
@@ -37,7 +37,7 @@ public class PlayerAuthRequest extends SendablePacket
 	}
 
 	protected void writeImpl(AuthServerClient client) {
-		writeByte(0x02);
+		writeByte((byte)0x02);
 		writeString(account);
 		writeInt(playOkID1);
 		writeInt(playOkID2);
